\name{praxis_pal}
\alias{praxis_pal}
\title{Return function to interpolate a praxis color palette}
\description{Return function to interpolate a praxis color palette}
\usage{
praxis_pal(palette = "main", reverse = FALSE, ...)}
\arguments{
  \item{palette}{Character name of palette in praxis_palettes}
  \item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
