\name{scale_color_praxis}
\alias{scale_color_praxis}
\title{Color scale constructor for praxis colors}
\usage{
scale_color_praxis(palette = "main", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in praxis_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for praxis colors
}
